Add Window Specs and Access to Pygame plus extras
Added class to aim_trainer_range to set window size and colors. Just
init def is designed so far. Plus a self for the clock.
aim_trainer_game added import for pygame as well as code to start and
stop it. Files added are sounds effects for shooting and targets.
Plus generic target image and fun font for prompts.