Integrate files into game.py and add backgrounds
game.py is coded to bring all files together to play game. Is not
functional yet, as some paths are not complete and will cause errors.
One to note is the difficulty selection. While the player is can be
prompted for the choice and the program can process a choice. There
is not a way to feed the program the choice made for processing.

Code and files were added to have a different background for each
stage of the game.